requirements.txt

annotated-types==0.7.0
appnope 
asgiref 
asttokens 
comm 
contextlib2==21.6.0
debugpy 
decorator 
Django 
django-countries
django-ninja==1.4.3
django-ninja-extra==0.30.1
et_xmlfile==2.0.0
exceptiongroup 
executing 
importlib_metadata 
injector==0.22.0
ipykernel 
ipython 
ipython_pygments_lexers 
jedi 
jupyter_client 
jupyter_core 
langchain==0.3.26
langgraph-checkpoint-postgres
langchain-community==0.3.27
langchain-core==0.3.74
langchain-groq==0.3.6
langchain-text-splitters==0.3.8
langgraph==0.6.6
langgraph-api==0.2.100
langgraph-checkpoint==2.1.1
langgraph-cli==0.3.5
langgraph-prebuilt==0.6.1
langgraph-runtime-inmem==0.6.1
langgraph-sdk==0.2.2
langsmith==0.4.8
matplotlib
matplotlib-inline 
nest_asyncio 
numpy==2.3.2
openpyxl==3.1.5
packaging 
pandas==2.3.2
parso 
pillow>=11.1.0,<11.2
psycopg2
psycopg[binary]
pexpect 
pickleshare 
platformdirs 
prompt_toolkit 
psutil 
ptyprocess 
pure_eval 
pydantic==2.11.7
pydantic_core==2.33.2
Pygments 
python-dateutil 
pytz==2025.2
pyzmq 
six 
SQLAlchemy==2.0.43
sqlparse 
stack_data 
tornado 
traitlets 
typing-inspection==0.4.1
typing_extensions 
tzdata==2025.2
uwsgi>=2.0.28,<2.1
wcwidth 
wordcloud
zipp 
